--- a/tests/performance/uniqExactIf.xml
+++ b/tests/performance/uniqExactIf.xml
@@ -1,0 +1,3 @@
+<test>
+    <query>SELECT uniqExactIf(number, ) FROM numbers_mt(1e7)</query>
+</test>