<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):

- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Parallel merge of `uniqExactIf` states. Closes #49885.

Before:
```sql
:) select uniqExactIf(number, 1) from numbers_mt(1e8);

SELECT uniqExactIf(number, 1)
FROM numbers_mt(100000000.)

Query id: 59752993-5e0a-45d6-afc9-1711cfde89a8

┌─uniqExactIf(number, 1)─┐
│              100000000 │
└────────────────────────┘

1 row in set. Elapsed: 7.744 sec. Processed 100.00 million rows, 800.00 MB (12.91 million rows/s., 103.31 MB/s.)
```

Now:
```sql
:) select uniqExactIf(number, 1) from numbers_mt(1e8);

SELECT uniqExactIf(number, 1)
FROM numbers_mt(100000000.)

Query id: fcc96bd1-10bd-401b-ba43-4ca20db9edbb

┌─uniqExactIf(number, 1)─┐
│              100000000 │
└────────────────────────┘

1 row in set. Elapsed: 1.291 sec. Processed 100.00 million rows, 800.00 MB (77.45 million rows/s., 619.59 MB/s.)
```