@Algunenano 2023-05-29
Does it make sense to enable this to other aggregate combinators too?