@ucasfl 2023-05-30
> Does it make sense to enable this to other aggregate combinators too?

I will do it in later PR.